+++ src/main/java/wooteco/subway/exception/ExceptionAdvice.java
@@ -6,13 +6,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import wooteco.subway.exception.addition.CannotAddException;
 import wooteco.subway.exception.auth.AuthorizationException;
-import wooteco.subway.exception.deletion.CannotDeleteException;
-import wooteco.subway.exception.duplication.DuplicatedException;
-import wooteco.subway.exception.line.InvalidDistanceException;
+import wooteco.subway.exception.badrequest.BadRequestException;
 import wooteco.subway.exception.notfound.NotFoundException;
-import wooteco.subway.exception.path.InvalidPathException;
 
 import java.sql.SQLException;
 import java.util.stream.Collectors;
@@ -31,10 +27,10 @@ public class ExceptionAdvice {
                 .body(message);
     }
 
-    @ExceptionHandler({CannotAddException.class, CannotDeleteException.class, DuplicatedException.class, InvalidPathException.class, InvalidDistanceException.class})
-    public ResponseEntity<String> handleBadRequest(RuntimeException runtimeException) {
+    @ExceptionHandler(BadRequestException.class)
+    public ResponseEntity<String> handleBadRequest(BadRequestException badRequestException) {
         return ResponseEntity.badRequest()
-                .body(runtimeException.getMessage());
+                .body(badRequestException.getMessage());
     }
 
     @ExceptionHandler(NotFoundException.class)
