--- src/main/java/wooteco/subway/exception/addition/CannotAddException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.exception.addition;
-
-public class CannotAddException extends RuntimeException {
-    public CannotAddException(String message) {
-        super(message);
-    }
-}
