+++ src/main/java/wooteco/subway/exception/auth/AuthorizationException.java
@@ -5,8 +5,6 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.UNAUTHORIZED)
 public class AuthorizationException extends RuntimeException {
-    public AuthorizationException() {
-    }
 
     public AuthorizationException(String message) {
         super(message);
