+++ src/main/java/wooteco/subway/exception/badrequest/BadRequestException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.exception.badrequest;
+
+public class BadRequestException extends RuntimeException {
+
+    public BadRequestException(String message) {
+        super(message);
+    }
+}
