+++ src/main/java/wooteco/subway/exception/badrequest/addition/CannotAddException.java
@@ -1,0 +1,10 @@
+package wooteco.subway.exception.badrequest.addition;
+
+import wooteco.subway.exception.badrequest.BadRequestException;
+
+public class CannotAddException extends BadRequestException {
+
+    public CannotAddException(String message) {
+        super(message);
+    }
+}
