+++ src/main/java/wooteco/subway/exception/badrequest/deletion/CannotDeleteException.java
@@ -1,0 +1,10 @@
+package wooteco.subway.exception.badrequest.deletion;
+
+import wooteco.subway.exception.badrequest.BadRequestException;
+
+public class CannotDeleteException extends BadRequestException {
+
+    public CannotDeleteException(String message) {
+        super(message);
+    }
+}
