+++ src/main/java/wooteco/subway/exception/badrequest/deletion/SectionCannotDeleteException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception.deletion;
+package wooteco.subway.exception.badrequest.deletion;
 
 public class SectionCannotDeleteException extends CannotDeleteException {
     private static final String MESSAGE = "노선에는 최소한 하나의 구간은 존재해야합니다.";
