+++ src/main/java/wooteco/subway/exception/badrequest/deletion/StationCannotDeleteException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception.deletion;
+package wooteco.subway.exception.badrequest.deletion;
 
 public class StationCannotDeleteException extends CannotDeleteException {
     private static final String MESSAGE = "이미 노선에 등록된 역은 삭제할 수 없습니다.";
