+++ src/main/java/wooteco/subway/exception/badrequest/duplication/DuplicatedException.java
@@ -1,0 +1,10 @@
+package wooteco.subway.exception.badrequest.duplication;
+
+import wooteco.subway.exception.badrequest.BadRequestException;
+
+public class DuplicatedException extends BadRequestException {
+
+    public DuplicatedException(String message) {
+        super(message);
+    }
+}
