+++ src/main/java/wooteco/subway/exception/badrequest/duplication/EmailDuplicatedException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception.duplication;
+package wooteco.subway.exception.badrequest.duplication;
 
 public class EmailDuplicatedException extends DuplicatedException {
     private static final String MESSAGE = "중복된 이메일입니다.";
