+++ src/main/java/wooteco/subway/exception/badrequest/duplication/LineColorDuplicatedException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception.duplication;
+package wooteco.subway.exception.badrequest.duplication;
 
 public class LineColorDuplicatedException extends DuplicatedException {
     private static final String MESSAGE = "잘못된 노선 색상입니다.";
