--- src/main/java/wooteco/subway/exception/deletion/CannotDeleteException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception.deletion;
-
-public class CannotDeleteException extends RuntimeException {
-
-    public CannotDeleteException(String message) {
-        super(message);
-    }
-}
