--- src/main/java/wooteco/subway/exception/duplication/DuplicatedException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception.duplication;
-
-public class DuplicatedException extends RuntimeException {
-
-    public DuplicatedException(String message) {
-        super(message);
-    }
-}
