+++ src/main/java/wooteco/subway/path/application/PathFinder.java
@@ -3,7 +3,7 @@ package wooteco.subway.path.application;
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.springframework.stereotype.Service;
-import wooteco.subway.exception.path.InvalidPathException;
+import wooteco.subway.exception.badrequest.path.InvalidPathException;
 import wooteco.subway.line.domain.Line;
 import wooteco.subway.path.domain.SectionEdge;
 import wooteco.subway.path.domain.SubwayGraph;
